+++ public/sw.js
@@ -1,0 +1,2 @@
+if(!self.define){let e,c={};const s=(s,a)=>(s=new URL(s+".js",a).href,c[s]||new Promise(c=>{if("document"in self){const e=document.createElement("script");e.src=s,e.onload=c,document.head.appendChild(e)}else e=s,importScripts(s),c()}).then(()=>{let e=c[s];if(!e)throw new Error(`Module ${s} didn’t register its module`);return e}));self.define=(a,i)=>{const n=e||("document"in self?document.currentScript.src:"")||location.href;if(c[n])return;let t={};const r=e=>s(e,n),d={module:{uri:n},exports:t,require:r};c[n]=Promise.all(a.map(e=>d[e]||r(e))).then(e=>(i(...e),t))}}define(["./workbox-4754cb34"],function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/KakaoIcon.png",revision:"050278218708b9197e0a468f0f5b2f3f"},{url:"/NaverIcon.png",revision:"9b8bbec2b446ff566ce2df35bc2d7905"},{url:"/_next/app-build-manifest.json",revision:"4ac3fa8be680b48904bced67288c7bb2"},{url:"/_next/static/TV6krNVy9-PwdfEcKFTMA/_buildManifest.js",revision:"a299a1a1b30054014daf38252a1526c7"},{url:"/_next/static/TV6krNVy9-PwdfEcKFTMA/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/1071-1c3c658fd2e4e875.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/1071-1c3c658fd2e4e875.js.map",revision:"0545ba1e71c9fc79f5dc1d292dd5575d"},{url:"/_next/static/chunks/137-dde80687dfeb7709.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/137-dde80687dfeb7709.js.map",revision:"2ddbd04d70c0fec74a81903583bf14b3"},{url:"/_next/static/chunks/162-807bcfaec2c769a9.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/162-807bcfaec2c769a9.js.map",revision:"ead1b551b801cfc23e701e2fcdb2e7b1"},{url:"/_next/static/chunks/1995-2e9dc2cb09b471b6.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/1995-2e9dc2cb09b471b6.js.map",revision:"700fdb26610ecfb5f787b8361cd81cc3"},{url:"/_next/static/chunks/2177-139a946c7ee564dc.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/2177-139a946c7ee564dc.js.map",revision:"3938fe7aabe39a0a4a5bacc963bcb466"},{url:"/_next/static/chunks/2319-cb52966cd03b1c7e.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/2319-cb52966cd03b1c7e.js.map",revision:"24116a72a1730897f4304e5d95773fc8"},{url:"/_next/static/chunks/2362-8476d04cfac3a5a5.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/2362-8476d04cfac3a5a5.js.map",revision:"d0446742e907543385b6f1fe0beec174"},{url:"/_next/static/chunks/2619-57d9df9f7ad67c51.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/2619-57d9df9f7ad67c51.js.map",revision:"d35e75aaef68b7459403a7d05db83e5d"},{url:"/_next/static/chunks/287-4b4ae6ba31e93e5f.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/287-4b4ae6ba31e93e5f.js.map",revision:"f18824608e21b1147a807f056ab5fe19"},{url:"/_next/static/chunks/2936-4c8b25d64069a3dc.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/2936-4c8b25d64069a3dc.js.map",revision:"c7612d07d24323aaefa3737b35183ca1"},{url:"/_next/static/chunks/295-a9d188ad5fceb9d0.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/295-a9d188ad5fceb9d0.js.map",revision:"428ae08c777aba394b802a10790d0901"},{url:"/_next/static/chunks/3539-7e67778c92547214.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/3539-7e67778c92547214.js.map",revision:"6ee61d766e4fd7345f8e3449893497b6"},{url:"/_next/static/chunks/3567-8ca3c0c9d601e64a.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/3567-8ca3c0c9d601e64a.js.map",revision:"03fceab1fe8782f22a293e13d84b4331"},{url:"/_next/static/chunks/4212-725b12e01ea206bd.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/4212-725b12e01ea206bd.js.map",revision:"c43cace874cacd176cbdda14a2466632"},{url:"/_next/static/chunks/472.2a11865d3e662869.js",revision:"2a11865d3e662869"},{url:"/_next/static/chunks/472.2a11865d3e662869.js.map",revision:"34848f06a95c698e62597d8037d6a3de"},{url:"/_next/static/chunks/4bd1b696-d3955faac4599738.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/4bd1b696-d3955faac4599738.js.map",revision:"e6aca4d17bc712e07d253297d2e44bed"},{url:"/_next/static/chunks/52774a7f-f0b8dd7cf3f34cf3.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/52774a7f-f0b8dd7cf3f34cf3.js.map",revision:"e0e93184d9e698765e89b8081b6f29c1"},{url:"/_next/static/chunks/5381-a937f65048ada192.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/5381-a937f65048ada192.js.map",revision:"e34185270353a42e8810be5efcc3ed6d"},{url:"/_next/static/chunks/5891-132440d9828b038a.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/5891-132440d9828b038a.js.map",revision:"0965fd6cbdf2b7e87269e30624b35118"},{url:"/_next/static/chunks/7359-e1a41127698e46f0.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/7359-e1a41127698e46f0.js.map",revision:"b41978aaa44e9c8818af78bd22f5503e"},{url:"/_next/static/chunks/778-2bb798ed6610a520.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/778-2bb798ed6610a520.js.map",revision:"28470e3c395ce83d8ef3444b5b76e88a"},{url:"/_next/static/chunks/802-c41dcb99e191e92d.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/802-c41dcb99e191e92d.js.map",revision:"e2c523c355822cbf6c274553ddf05788"},{url:"/_next/static/chunks/8192-0755e76528d1769e.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/8192-0755e76528d1769e.js.map",revision:"ba8128e5cca9e4f650e0da792d657302"},{url:"/_next/static/chunks/8220-631d20ae93272d3a.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/8220-631d20ae93272d3a.js.map",revision:"dfdea9935baf240a771b93b324a45410"},{url:"/_next/static/chunks/8633-5b63e4bf611512c9.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/8633-5b63e4bf611512c9.js.map",revision:"0bf470a73c330e62faed2868d7505376"},{url:"/_next/static/chunks/881-d6aef57ae8571833.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/881-d6aef57ae8571833.js.map",revision:"791d05afb780f60a51dffcc9b475d463"},{url:"/_next/static/chunks/900-1854833c22b78b3e.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/900-1854833c22b78b3e.js.map",revision:"33526a4e2caf7d9aa13bca2020ccf041"},{url:"/_next/static/chunks/9088-a52b9652f1098db0.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/9088-a52b9652f1098db0.js.map",revision:"3d99bcf4306c4b8189641f1327c6f978"},{url:"/_next/static/chunks/9156-941c22162a3f78a2.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/9156-941c22162a3f78a2.js.map",revision:"b361e162f4a44451b77b1a1ee38c73c1"},{url:"/_next/static/chunks/9160-4a392f5d4bdad300.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/9160-4a392f5d4bdad300.js.map",revision:"53faf1c6822204406ae40567e1355069"},{url:"/_next/static/chunks/9341.31dbdebe0cb82e35.js",revision:"31dbdebe0cb82e35"},{url:"/_next/static/chunks/9341.31dbdebe0cb82e35.js.map",revision:"46d552b29c246fa4475df8b3d39fcf03"},{url:"/_next/static/chunks/9415-08911124683ee758.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/9415-08911124683ee758.js.map",revision:"854552e46c7a89b88d09560d2a71091d"},{url:"/_next/static/chunks/9519-8bd199151c4efdcc.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/9519-8bd199151c4efdcc.js.map",revision:"30d5b01b540826dd22ab7ad2f5278ce5"},{url:"/_next/static/chunks/9666-22bf864d89615ab3.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/9666-22bf864d89615ab3.js.map",revision:"2a7cbeccfa2a7bbdd57f646878eb0e5b"},{url:"/_next/static/chunks/9761-a4446ede8e13c50b.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/9761-a4446ede8e13c50b.js.map",revision:"fdd6e4964e096fca27500e41417d68e4"},{url:"/_next/static/chunks/app/_not-found/page-b48fa8b2b5d19532.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/app/_not-found/page-b48fa8b2b5d19532.js.map",revision:"8502519fd448fc3a422e0085d563cec2"},{url:"/_next/static/chunks/app/api/auth/refresh/route-5297619cd5220076.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/app/api/sentry-example-api/route-931170d056c9667e.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/app/calendar/%5Bdate%5D/page-2233c726dc47cf9d.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/app/calendar/%5Bdate%5D/page-2233c726dc47cf9d.js.map",revision:"d8f59e80fd1c27af67628359ecd20513"},{url:"/_next/static/chunks/app/global-error-34de649ab7b5dd45.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/app/global-error-34de649ab7b5dd45.js.map",revision:"d3cc3abcf59ec6b0107d83204f261cae"},{url:"/_next/static/chunks/app/ingredients/new/page-a34c5af54ed41268.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/app/ingredients/new/page-a34c5af54ed41268.js.map",revision:"608125acaeb6f0b0a8fd5ece984b46bb"},{url:"/_next/static/chunks/app/ingredients/page-8139b9fa7d3604d2.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/app/ingredients/page-8139b9fa7d3604d2.js.map",revision:"c2d1e6b12f13c62fb2ef29534d0197e5"},{url:"/_next/static/chunks/app/layout-3ba35a4e7c7d5413.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/app/layout-3ba35a4e7c7d5413.js.map",revision:"5dffdf2c9f28a15153a2f4dfc243d968"},{url:"/_next/static/chunks/app/login/layout-edfbb6f1dc02f633.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/app/login/page-24e684e75e28a90a.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/app/login/page-24e684e75e28a90a.js.map",revision:"1e812be4940768e8257ddcfec9364eb3"},{url:"/_next/static/chunks/app/page-68288da2759283dd.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/app/page-68288da2759283dd.js.map",revision:"54702ec66a1e743a992dd4e5675ef5bf"},{url:"/_next/static/chunks/app/recipes/%5BrecipeId%5D/comments/%5BcommentId%5D/page-32c41916879f8530.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/app/recipes/%5BrecipeId%5D/comments/%5BcommentId%5D/page-32c41916879f8530.js.map",revision:"dece1bc293c2469adb9bf6f1fa142cf9"},{url:"/_next/static/chunks/app/recipes/%5BrecipeId%5D/comments/page-c97f4f2df0dfb2c5.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/app/recipes/%5BrecipeId%5D/comments/page-c97f4f2df0dfb2c5.js.map",revision:"e333e8216921395dadb1ffafbcf61f2c"},{url:"/_next/static/chunks/app/recipes/%5BrecipeId%5D/edit/page-a316b24fc86d44dd.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/app/recipes/%5BrecipeId%5D/edit/page-a316b24fc86d44dd.js.map",revision:"2f849c86abb1b278ec211f6bc886a1d3"},{url:"/_next/static/chunks/app/recipes/%5BrecipeId%5D/layout-8fe66bfa68a0e6b0.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/app/recipes/%5BrecipeId%5D/page-d56b6db3a670c57f.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/app/recipes/%5BrecipeId%5D/page-d56b6db3a670c57f.js.map",revision:"c7ebf731c60eea4eeb1d270fc509e30a"},{url:"/_next/static/chunks/app/recipes/%5BrecipeId%5D/rate/page-f5c004a9aeaee21f.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/app/recipes/%5BrecipeId%5D/rate/page-f5c004a9aeaee21f.js.map",revision:"f20b46923076b7be35ddbb43bd1f055e"},{url:"/_next/static/chunks/app/recipes/%5BrecipeId%5D/slide-show/layout-141e1327de272a9a.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/app/recipes/%5BrecipeId%5D/slide-show/page-cd75a606f4143b85.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/app/recipes/%5BrecipeId%5D/slide-show/page-cd75a606f4143b85.js.map",revision:"3884c052519c25cce04fd716a807fc84"},{url:"/_next/static/chunks/app/recipes/ai/page-8adb7398de314094.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/app/recipes/ai/page-8adb7398de314094.js.map",revision:"ba0cf5b2e664bdd5a5346c422cc2d195"},{url:"/_next/static/chunks/app/recipes/category/%5Bid%5D/page-976e949d19e5129b.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/app/recipes/category/%5Bid%5D/page-976e949d19e5129b.js.map",revision:"b9d68722ba7270abb000248cd005605b"},{url:"/_next/static/chunks/app/recipes/new/ai/page-695cbb9a68741a60.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/app/recipes/new/ai/page-695cbb9a68741a60.js.map",revision:"8990942c4c642d795d34a06c90636772"},{url:"/_next/static/chunks/app/recipes/new/page-9d121513d4ccf40c.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/app/recipes/new/page-9d121513d4ccf40c.js.map",revision:"fe1c8072f4c9a410b5fa6b9fa5a325b2"},{url:"/_next/static/chunks/app/search/page-2ad9bca55f9b4af6.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/app/search/page-2ad9bca55f9b4af6.js.map",revision:"8796599227b3c17ae4513c9a9cb1654c"},{url:"/_next/static/chunks/app/sentry-example-page/page-cbf11517f9efc30a.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/app/sentry-example-page/page-cbf11517f9efc30a.js.map",revision:"f8ca42c6850dfede9a642544caf53834"},{url:"/_next/static/chunks/app/users/%5BuserId%5D/page-e896e7499377cfd7.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/app/users/%5BuserId%5D/page-e896e7499377cfd7.js.map",revision:"473442b0392455a14bc07c79ec757241"},{url:"/_next/static/chunks/app/users/edit/page-0a403d193fbb9036.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/app/users/edit/page-0a403d193fbb9036.js.map",revision:"79a0d34876d1fbb39ee30bd61b017924"},{url:"/_next/static/chunks/c15bf2b0-3b74866c582ad21a.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/c15bf2b0-3b74866c582ad21a.js.map",revision:"03fd2c3462d41fb590854f905cc49c91"},{url:"/_next/static/chunks/framework-4722183a3be1a5d4.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/framework-4722183a3be1a5d4.js.map",revision:"cd420a015d3ff3ab32f2c47717a594e9"},{url:"/_next/static/chunks/main-40b833e239ff6c86.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/main-40b833e239ff6c86.js.map",revision:"cf5a4c2331815c0144d7e23958a9d969"},{url:"/_next/static/chunks/main-app-b2d09a6eeacd1be3.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/main-app-b2d09a6eeacd1be3.js.map",revision:"064a356b263d929c1fe5dede2838d15f"},{url:"/_next/static/chunks/pages/_app-315cf73dd3558e85.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/pages/_app-315cf73dd3558e85.js.map",revision:"a7901641290e51a11369e661fc158766"},{url:"/_next/static/chunks/pages/_error-6b0a99f11bb0d686.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/pages/_error-6b0a99f11bb0d686.js.map",revision:"2b4c11948cc4497dfc3486773dd18c2e"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-ec8ea4298fee9ffb.js",revision:"TV6krNVy9-PwdfEcKFTMA"},{url:"/_next/static/chunks/webpack-ec8ea4298fee9ffb.js.map",revision:"2cfd1b73ae1a3ce883d9ec6849f99bdd"},{url:"/_next/static/css/9beac7c2ffd5c91d.css",revision:"9beac7c2ffd5c91d"},{url:"/_next/static/css/9beac7c2ffd5c91d.css.map",revision:"bcf5e0536ed88bbce1d42aa859080ceb"},{url:"/_next/static/css/c66e46f533e6f7b8.css",revision:"c66e46f533e6f7b8"},{url:"/_next/static/css/c66e46f533e6f7b8.css.map",revision:"c8d0f9ad2b3dcc986ba6c7718dc0186c"},{url:"/logo.svg",revision:"703fd94fee4972ae176026390d422d8f"},{url:"/manifest.json",revision:"6688c497d11734682a52b0f2d6c6cb14"},{url:"/pwa_logo.png",revision:"7f2e04c2fa0c0fac3cf75647801dff78"},{url:"/robot1.png",revision:"2eb4a4f42b4b659c7d2030c7cda15116"},{url:"/screenshots/desktop-1.png",revision:"7f2e04c2fa0c0fac3cf75647801dff78"},{url:"/screenshots/mobile-1.png",revision:"7f2e04c2fa0c0fac3cf75647801dff78"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:c,event:s,state:a})=>c&&"opaqueredirect"===c.type?new Response(c.body,{status:200,statusText:"OK",headers:c.headers}):c}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(({url:e})=>{if(!(self.origin===e.origin))return!1;const c=e.pathname;return!c.startsWith("/api/auth/")&&!!c.startsWith("/api/")},new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")},new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(({url:e})=>!(self.origin===e.origin),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")});
+//# sourceMappingURL=sw.js.map
